refactor(core-run): split key-provider into a single file
